+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+### 35.8.0 [#1114](https://github.com/openfisca/openfisca-core/pull/1114)
+
+#### New Features
+
+- Introduce `rate_from_bracket_indice` method on `RateTaxScaleLike` class
+  - Allows for the determination of the tax rate based on the tax bracket indice
+
+- Introduce `rate_from_tax_base` method on `RateTaxScaleLike` class
+  - Allows for the determination of the tax rate based on the tax base
+
+- Introduce `threshold_from_tax_base` method on `RateTaxScaleLike` class
+  - Allows for the determination of the lower threshold based on the tax base
+
 ### 35.7.8 [#1086](https://github.com/openfisca/openfisca-core/pull/1086)
 
 #### Technical changes